comment增删时操作对应article

+++ src/main/java/edu/xidian/pixels/Service/CommentService.java
@@ -2,6 +2,7 @@ package edu.xidian.pixels.Service;
 
 import edu.xidian.pixels.Entity.Article;
 import edu.xidian.pixels.Entity.Comment;
+import edu.xidian.pixels.Mapper.ArticleMapper;
 import edu.xidian.pixels.Mapper.CommentMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,6 +14,9 @@ public class CommentService {
 
     @Autowired
     private CommentMapper commentMapper;
+
+    @Autowired
+    private ArticleMapper articleMapper;
 
     public Comment findById(Integer id){
         if(id!=null){
@@ -42,17 +46,32 @@ public class CommentService {
     }
 
     public boolean insert(Comment comment){
-        if(comment!=null){
-            if(commentMapper.insert(comment)>0)
-                return true;
+        if(comment!=null) {
+            Article article = articleMapper.findById(comment.getArticleId());
+            if (article != null) {
+                article.setComments(article.getComments()+1);
+                if (commentMapper.insert(comment) > 0) {
+                    articleMapper.update(article);
+                    return true;
+                }
+            }
         }
         return false;
     }
 
     public boolean delete(Integer id){
         if(id!=null){
-            if(commentMapper.delete(id)==1)
-                return true;
+            Comment comment=commentMapper.select(id);
+            if(comment!=null){
+                Article article=articleMapper.findById(comment.getArticleId());
+                if(article!=null){
+                    article.setComments(article.getComments()-1);
+                    if(commentMapper.delete(id)==1) {
+                        articleMapper.update(article);
+                        return true;
+                    }
+                }
+            }
         }
         return false;
     }
